move a couple of files
