+++ package.json
@@ -7,15 +7,15 @@
     "node": ">= 14.16 <15",
     "npm": "^6.12.0"
   },
-  "main": "src/main",
+  "main": "src/backend/main",
   "scripts": {
     "prestart": "ts-node env.ts",
     "start": "run-p start:main start:worker",
-    "start:main": "ts-node-dev --inspect=0.0.0.0:9229 --respawn --transpile-only --rs src/main",
-    "start:worker": "ts-node-dev --inspect=0.0.0.0:9230 --respawn --transpile-only --rs src/worker",
+    "start:main": "ts-node-dev --inspect=0.0.0.0:9229 --respawn --transpile-only --rs src/backend/main",
+    "start:worker": "ts-node-dev --inspect=0.0.0.0:9230 --respawn --transpile-only --rs src/backend/worker",
     "start:production": "run-p start:main:production start:worker:production",
     "start:main:production": "run-s db:migrate:prod && node .",
-    "start:worker:production": "run-s db:migrate:prod && node src/worker",
+    "start:worker:production": "run-s db:migrate:prod && node src/backend/worker",
     "clean": "rimraf coverage build tmp src/*.js src/*.js.map",
     "build": "tsc -p ./tsconfig.json",
     "build:release": "tsc -p ./tsconfig.release.json",
