+++ src/backend/configure-robot.ts
@@ -4,17 +4,17 @@ import RateLimit from "express-rate-limit";
 import RedisStore from "rate-limit-redis";
 import { createAppAuth } from "@octokit/auth-app";
 import { findPrivateKey } from "probot/lib/private-key";
-import getRedisInfo from "./config/redis-info";
-import setupFrontend from "./frontend";
-import setupGitHub from "./github";
-import statsd from "./config/statsd";
-import { isIp4InCidrs } from "./config/cidr-validator";
+import getRedisInfo from "../config/redis-info";
+import setupFrontend from "../frontend";
+import setupGitHub from "../github";
+import statsd from "../config/statsd";
+import { isIp4InCidrs } from "../config/cidr-validator";
 import Logger from "bunyan";
-import { metricError } from "./config/metric-names";
-import logMiddleware from "./middleware/log-middleware";
+import { metricError } from "../config/metric-names";
+import logMiddleware from "../middleware/log-middleware";
 
 
-import { getLogger } from "./config/logger";
+import { getLogger } from "../config/logger";
 
 
 /**
